+++ src/main/types.ts
@@ -353,6 +353,15 @@ enum StorageFilter {
 }
 
 /**
+ * Lock filtering options.
+ */
+enum LockFilter {
+  ALL = 'all',
+  LOCKED = 'locked',
+  UNLOCKED = 'unlocked',
+}
+
+/**
  * The state of the frontend.
  */
 type AppState = {
@@ -364,6 +373,7 @@ type AppState = {
   videoFilterTags: Tag[];
   dateRangeFilter: DateValueType;
   storageFilter: StorageFilter;
+  lockFilter: LockFilter;
   videoFullScreen: boolean;
   playing: boolean;
   language: Language;
@@ -760,6 +770,7 @@ export {
   CreateMultiPartUploadResponseBody,
   CompleteMultiPartUploadRequestBody,
   StorageFilter,
+  LockFilter,
   ObsSourceCallbackInfo,
   ObsVolmeterCallbackInfo,
   VideoSourceName,