+++ src/renderer/VideoPlayer.tsx
@@ -2,6 +2,7 @@ import {
   AppState,
   DeathMarkers,
   InstantReplayData,
+  LockFilter,
   RendererVideo,
   SliderMark,
   StorageFilter,
@@ -118,8 +119,16 @@ export const VideoPlayer = forwardRef<VideoPlayerRef, IProps>((props, ref) => {
     instantReplay,
   } = props;
 
-  const { playing, multiPlayerMode, language, selectedVideos, storageFilter } =
-    appState;
+  const {
+    playing,
+    multiPlayerMode,
+    language,
+    selectedVideos,
+    storageFilter,
+    lockFilter,
+  } = appState;
+  const hasActiveFilter =
+    storageFilter !== StorageFilter.BOTH || lockFilter !== LockFilter.ALL;
 
   if (!instantReplay && (videos.length < 1 || videos.length > 4)) {
     // Protect against stupid programmer errors.
@@ -911,7 +920,7 @@ export const VideoPlayer = forwardRef<VideoPlayerRef, IProps>((props, ref) => {
    * Render the download button.
    */
   const renderDownloadButton = () => {
-    const disabled = storageFilter !== StorageFilter.BOTH;
+    const disabled = hasActiveFilter;
     const tooltip = disabled
       ? getLocalePhrase(language, Phrase.DisabledDueToFilter)
       : getLocalePhrase(language, Phrase.DownloadButtonTooltip);
@@ -936,7 +945,7 @@ export const VideoPlayer = forwardRef<VideoPlayerRef, IProps>((props, ref) => {
    * Render the upload button.
    */
   const renderUploadButton = () => {
-    const disabled = storageFilter !== StorageFilter.BOTH;
+    const disabled = hasActiveFilter;
     const tooltip = disabled
       ? getLocalePhrase(language, Phrase.DisabledDueToFilter)
       : getLocalePhrase(language, Phrase.UploadButtonTooltip);